+++ src/pages/memberManagement/MemberInfoPage.tsx
@@ -2,8 +2,8 @@ import MemberInfoBasicTable from '@/components/MemberInfo/MemberBasicInfoTable';
 
 export default function MemberInfoPage() {
   return (
-    <>
+    <div className="h-min-screen">
       <MemberInfoBasicTable />
-    </>
+    </div>
   );
 }
